--- setup.py
@@ -1,6 +1,0 @@
-from setuptools import find_packages, setup 
-
-setup(
-    name='petdeface', 
-    packages=find_packages(),
-)
